Add support for custom errors
This commit adds support for custom errors. The new logic allows
for setting additional values inside the JS error object,
compared to Errors exposed by the napi-rs, that allows only setting
error message and kind.

This is intended to be a easy replacement
of the current errors with only minor refactor required.
You should only need to wrap the body of the function into
with_custom_error_(sync/async), and update the return type from
napi::Result<T> to JsResult<T>.

Example of such change is added in a following commit.

Co-authored-by: Wojciech Przytuła <wojciech.przytula@scylladb.com>