Raw results
This commit adds an ability to get the following page of the result,
without wrapping the results into result set. This is done,
by updating method on the ResultSet class, `rawNextPageAsync()` from `nextPageAsync()`,
that returns the page state and result wrappers, that would otherwise
be used to construct the next ResultSet. This can be done as nextPageAsync
is only internal method.

This optimization is then used in the result async iterator,
which already only takes the page state and rows from the result sets
for the following pages. This change optimizes async iteration over result rows.

This commit removes some of the unit tests, as they depend on the
old implementation details of the `nextPageAsync()`.