Move results page decoding to JS
This commit updates the deserialization process from the Rust layer,
to the JavaScript code. Compared to the previous attempt,
here we pass the whole page of results, instead of deserializing
individual values.

This leans to the better performance, as we do no create individual
buffers for each values, through the Node-API layer, and only create
subslice of the page buffer, when deserializing values in JS.

FrameReader is a heavily reduced version of the DataStax implementation.
This reader was not tested in the DSx driver, and I introduce unit tests
for this class in the following commit.