Update queried host flakey test (#362)
The integration test introduced in #357 turned out to be flakey when
testing against Scylla.

For Scylla, we may get the address with shard aware port,
while the client.hosts returns the hosts with non shard aware ports.
(See https://github.com/scylladb/scylla-rust-driver/issues/1513)

For this reason I updated the test, when testing with Scylla,
to strip the port part of the address, fixing the flakiness.