Update cargo toml
This updates two things in the cargo.toml:
1. Panic mode in release mode. See #47 for motivation of this change.
2. Remove strip flag from the release mode. Why:
This flag was part of the napi auto-generated template. The binary built
without this flag is about ~15% bigger (4.4MB -> 5MB). This increase
allows a better collection of crash logs (as we obtain full stack trace)
during panics that are meant to indicate driver bugs.

This change also matches the rust drier (which does not strip symbols)