Update handling of JS time
This change removes the Rust inner value, to keep the value
just on the JS side. This change visibly increases the performance
of the driver for the deser benchmarks.

The new version is based on the DSx implementation, with the following
changes:
- Increased validation in construction. The previous version (implemented
in rust) had strict checks on validity - I add those checks to JS code
- Read only: to match the DSx description of the class
(and the previous implementation), I make all the fields read only.

Performance impact:
About 15-20% execution time improvement on concurrent deser
(see PR description for more detailed numbers).