Update handling of Js Inet
This change removes the Rust inner value, to keep the value
just on the JS side. This change visibly increases the performance
of the driver for the deser benchmarks.

The new version is based on the DSx implementation, with the changes
already introduced, when working initially on this type

Performance impact:
About 20% execution time improvement on concurrent deser
(see PR description for more detailed numbers).