Update default encoder bigInt options
This commit updates the default values for encoding useBigIntAsLong
and useBigIntAsVarint client options. With this change, the default
values encourage usage of builtin type (BigInt), instead of
external type (Long) or type with custom implementation (integer.js).

See #346 for more details.

For the integration tests that are broken by the change of the default values,
I update them to manually override those options to the old default values.
For the integration tests that were requiring setting those options to new defaults,
I update them to rely on the default values, instead of overriding them.